requirements.txt
abstract==2020.7.26
asgiref==3.3.1
auth==0.5.3
auth-token==0.0.3
base32hex==1.0.2
blinker==1.4
certifi==2020.4.5.2
chardet==3.0.4
colouration==2020.7.26
defusedxml==0.6.0
Django==3.1.6
django-bigautohack==0.1
django-cors-headers==3.4.0
django-multiselectfield==0.1.12
django-redis==4.12.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.8.0
dnspython==1.16.0
docopt==0.6.2
eventlet==0.31.1
falcon==3.0.1
filemagic==1.6
graphviz==0.16
greenlet==1.1.1
gunicorn==20.1.0
idna==2.9
mongoengine==0.23.1
oauthlib==3.1.0
Pillow==8.1.0
pipreqs==0.4.10
PyJWT==1.7.1
pymongo==3.12.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2021.1
redis==3.5.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
urllib3-1-26-2==1.26.2.0
Werkzeug==2.0.1
yarg==0.1.9
